FEAT: add type_bmi and fix models

+++ health_infos/models.py
@@ -7,5 +7,5 @@ class Heath_Info(models.Model):
     id = models.UUIDField(primary_key=True, default=uuid4, unique=True)
     height = models.DecimalField(max_digits=3, decimal_places=2)
     weight = models.DecimalField(max_digits=8, decimal_places=2)
-    bmi = models.IntegerField()
+    bmi = models.DecimalField(max_digits=4, decimal_places=2)
     ideal_weight = models.DecimalField(max_digits=8, decimal_places=2)
